+++ src/embeds/skills.js
@@ -1,7 +1,9 @@
 // @flow
 
+import type { EmbedProps } from './bootstrap';
+
 import Skills from './components/Skills';
 
 export default function (element: HTMLElement, ids: Array<number>) {
-  return () => <Skills ids={ids} />;
+  return (props: EmbedProps) => <Skills {...props} ids={ids} />;
 }