+++ CHANGELOG.md
@@ -8,6 +8,14 @@
 - Changes
 ```
 
+## [4.3.0] - 2017/03/11
+### Added
+- Ability to turn off embed badge
+- Classes for consumers of embeds to use to override styles
+
+### Changed
+- Armory badge for tooltip to have dark background
+
 ## [4.2.1] - 2017/03/09
 ### Fixed
 - i18n being set with `'undefined'`